move styles sheests to /src/styles
