+++ src/styles/ActivePlayer.css
@@ -4,7 +4,7 @@
     position: absolute;
     top: 0%;
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
     justify-content: center;
     align-items: center;
 }
